# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API schema",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Просмотр всех пользователей в системе с опциональной пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Список пользователей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Размер страницы с пользователями",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страницы с пользователями",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_users_transport_http.UserDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal servver error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "CreateUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Получение конкретного пользователя по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID получаемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно найден",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление существующего пользователя по его ID",
                "tags": [
                    "users"
                ],
                "summary": "Удаление пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID удаляемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление пользователя"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменение данных об уже существующем в системе пользователе\n### Логика обновления полей (Three-state logic):\n1. **Поле не передано**: `phone_number` игнорируется, значение в БД не меняется\n2. **Явно передано значение** : `\"phone_number\": \"+71112223344\"` - устанавливает новый номер телефона в БД\n3. **Передан null**: `phone_number\": null` - очищает поле в БД (set to NULL)\nОграничения: `full_name` не может быть выставлен как null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Изменение пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изменяемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.PatchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно измененный пользователь",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.PatchUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_moond0wner_todo-nilchan_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
...
        "internal_features_users_transport_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Ivanov Ivan"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10,
                    "example": "+79998887766"
                }
            }
        },
        "internal_features_users_transport_http.CreateUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_users_transport_http.GetUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_users_transport_http.PatchUserRequest": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Максим Максимович"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                }
            }
        },
        "internal_features_users_transport_http.PatchUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_users_transport_http.UserDTOResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        }
    }
}